fix(quiz): mengubah tipe data nilai kuis menjadi integer dan menghilangkan pembulatannya pada page

+++ database/migrations/2025_07_27_114927_create_user_quizzes_table.php
@@ -27,7 +27,7 @@ return new class extends Migration
                     ->on('submodule_quizzes')
                     ->onDelete('cascade');
             $table->integer('attempt_count')->default(0);
-            $table->float('high_score')->default(0);
+            $table->integer('high_score')->default(0);
             $table->timestamps();
         });
     }
